Fix dev:setup scripts
It was noticed that nested symlinks were not being set up correctly based
on the fact that the utils templates did not fail tests when changed.
It turns out that there is a very specific order in which symlinks need
to be created in. I have triple-checked this and have confirmed that not
only do global links for @cremalab/scripts-mobile|web are created, but
each of them has a symbolic link to the base @cremalab/scripts package.