Update app-mobile new:component output structure
Moves away from pattern where main implementation was contained within
the `A/index.ts` file in favor of implementation within a named
module (e.g. `A.ts`). This enables better DX because it is much easier
to search for the module by name. The test file has been renamed for the
same purposes. The final directory structure looks like this:

```
.
└── components
    └── A
        ├── A.stories.tsx
        ├── A.test.tsx
        ├── A.tsx
        ├── README.md
        └── index.ts
```