Change module structure (#150)
* Update new:util output structure

Moves away from pattern where main implementation was contained within
the `foo/index.ts` file in favor of implementation within a named module
(e.g. `foo/foo.ts`). This enables better DX because it is much easier to
search for the module by name. The test file has been renamed for the
same purposes. The final directory structure looks like this:

```
.
└── utils
    └── foo
        ├── foo.test.ts
        ├── foo.ts
        └── index.ts
```

* Update app-web new:component output structure

Moves away from pattern where main implementation was contained within
the `A/index.ts` file in favor of implementation within a named
module (e.g. `A.ts`). This enables better DX because it is much easier
to search for the module by name. The test file has been renamed for the
same purposes. The final directory structure looks like this:

```
.
└── components
    └── A
        ├── A.stories.tsx
        ├── A.test.tsx
        ├── A.tsx
        ├── README.md
        └── index.ts
```

* Remove babel-jest since unused

* chore(deps): update dependency ts-jest to v26.5.5

* Fix generateNewComponent inconsistencies

* Update app-mobile new:component output structure

Moves away from pattern where main implementation was contained within
the `A/index.ts` file in favor of implementation within a named
module (e.g. `A.ts`). This enables better DX because it is much easier
to search for the module by name. The test file has been renamed for the
same purposes. The final directory structure looks like this:

```
.
└── components
    └── A
        ├── A.stories.tsx
        ├── A.test.tsx
        ├── A.tsx
        ├── README.md
        └── index.ts
```

Co-authored-by: Renovate Bot <bot@renovateapp.com>