GH-157: Add `sh: npx eslint --fix` to templates
This script is added to the last template file for each generator since
it needs to run after all files have been created. Hygen doesn't provide
a way to determine when the shell script is executed, so this is the best
solution I found since Hygen runs each template file in alphanumeric order.