GH-157: Run formatter on generated code (#162)
* GH-157: Add `sh: npx eslint --fix` to templates

This script is added to the last template file for each generator since
it needs to run after all files have been created. Hygen doesn't provide
a way to determine when the shell script is executed, so this is the best
solution I found since Hygen runs each template file in alphanumeric order.

* GH-157: Update scripts-web component templates

Now that we are linting and fixing files upon generation, it became clear
that our "with examples" variants produced code with errors. Beyond this,
some of the example code was confusing. Changes include:
- Define `Props` as an `interface` with a `name` property
- Destructure `props` to access `name` parameter (very common)
- Update tests to ensure that the `name` prop is rendered

* GH-157: Update scripts-mobile component templates

Now that we are linting and fixing files upon generation, it became clear
that our "with examples" variants produced code with errors. Beyond this,
some of the example code was confusing. Changes include:
- Define `Props` as an `interface` with a `name` property
- Destructure `props` to access `name` parameter (very common)
- Update tests to ensure that the `name` prop is rendered