GH-180: Trim generator input (#181)
* GH-180: Trim generator input - new:type

* GH-180: Trim generator input - new:util

* GH-180: Add bootstrap to beginning

Lerna bootstrap installs all package dependencies. Now that `inflection`
is imported in a TypeScript file, it has uncovered that this dependency
wasn't ever present up-front but instead was present at the end. Now TS
requires tht it be installed before building—of course!

* GH-180: Use Node 16.x in workflow

* GH-180: Update scripts package-lock.json to v2

* GH-180: Fix incorrect binary reference in scripts-mobile

* GH-180: Add workflow step to update npm

* GH-180: Update actions/setup-node to v2

* GH-180: Update app-mobile submodule ref

* GH-180: Lock cross-spawn across project

* GH-180: Add -e flag to echo for embedded new lines

* GH-180: Print Node/NPM version at consumer install

* GH-180: Remove sudo from ci commands

This was causing consumers to run on the system version of node/npm
which caused symlink references to be incorrect. We now print these
versions in the consumer context to allow for better debugging in
the future.

* GH-180: Increase jest timeout to 1 minute

These tests run faster/slower depending on the environment. When using
[act](https://github.com/nektos/act) to test actions locally, they were
timing out. I'm willing to wait a minute to ensure they _really_ fail.

* GH-180: Trim generator input - new:component

* GH-180: Downgrade to node@15